Formalise external event API
This commit brings the event emitter returned by the API up to the top
level API methods to make it simpler to control the API level events.
Previously it was too easy to mistakenly alter the external API from
deep inside the library.

In addition, the external event API is altered.

For backup:

- Rename `writeerror` event as just `error`.
- Rename `writecomplete` event to `finished`.

Plain `error` seems to be the convention in the Node.js std lib.
I chose `finished` because it doesn't imply success in the same way
that `complete` does.

Restore:

- Rename `writeerror` event as just `error`.
- Rename `writecomplete` event to `finished`.
- Rename `written` restore event `restored`

`restored` feels a bit more domain relevant, and the std libs tend
towards domain relevance when events are not really generic, like
`error`.

#36