fix: wrap ERR_INVALID_URL in InvalidOption
Export wrapPossibleInvalidUrlError from error module that checks for ERR_INVALID_URL
and wraps in InvalidOption.
Call wrapPossibleInvalidUrlError from app.js and cliutils databaseUrl to ensure that
both places that construct the URL return the correct error type.
Add try/catch blocks to couchbackup.bin.js and couchrestore.bin.js that call the
error.terminationCallback so that we get correct exit codes and not unhandled errors
if an error occurs before we are in the backup or restore functions.
Remove TypeError from test assertions as it should have the correct InvalidOption type
now.